+++ app/src/main/res/menu/menu_main.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+    <item
+        android:id="@+id/menuitem_register"
+        android:title="Registrar" />
+    <item
+        android:id="@+id/menuitem_salir"
+        android:title="Salir" />
+</menu>
