+++ app/src/main/res/values/colors.xml
@@ -2,4 +2,13 @@
 <resources>
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
+
+    <!-- Custom Colors -->
+    <color name="blue">#337dff</color>
+    <color name="grey">#e7e8eb</color>
+    <color name="grisclaro">#E0E0E0</color>
+    <color name="red">#E61212</color>
+    <color name="green">#4CF013</color>
+    <color name="celesteclaro">#F5FAFF</color>
+    <color name="grisoscuro">#6D6D6D</color>
 </resources>
